wanpipe-3.1.0.tgz

+++ patches/kdrivers/include/sdla_xilinx.h
@@ -1040,7 +1040,7 @@ typedef struct {
 
 typedef struct wp_rx_element
 {
-	unsigned long dma_addr;
+	unsigned int dma_addr;
 	unsigned int reg;
 	unsigned int align;
 	unsigned char pkt_error;
